admin tool

+++ users/views.py
@@ -68,9 +68,35 @@ def logout_view(request):
 
 @login_required
 def clear_files(request):
-    # Only allow superusers to clear files
+    # Only allow superusers to access page
     if request.user.is_superuser:
-        # Delete all files in the music directory
-        if Path(settings.MUSIC_DIR).exists():
-            rmtree(settings.MUSIC_DIR)
+        if request.method == "PUT":
+            # Delete all files in the music directory
+            if Path(settings.MUSIC_DIR).exists():
+                rmtree(settings.MUSIC_DIR)
+        else:
+            files = []
+            for folder in Path(settings.MUSIC_DIR).iterdir():
+                for file in folder.iterdir():
+                    if folder.name not in ["Playlists", "Tracks"]:
+                        files.append(file.name)
+
+            playlists = []
+            for folder in Path.joinpath(settings.MUSIC_DIR, "Playlists").iterdir():
+                for playlist in folder.iterdir():
+                    playlists.append(playlist.name)
+            tracks = []
+            for folder in Path.joinpath(settings.MUSIC_DIR, "Tracks").iterdir():
+                for track in folder.iterdir():
+                    tracks.append(track.name)
+
+            return render(
+                request,
+                "users/clear_files.html",
+                {
+                    "files": files,
+                    "playlists": playlists,
+                    "tracks": tracks,
+                },
+            )
     return HttpResponseRedirect(reverse("index"))
